[BMSPT-174] reformat BcfBuilder
