+++ src/bcf-toolkit/Builder/Bcf21/BcfBuilder.cs
@@ -1,5 +1,4 @@
 using System;
-using System.Collections.Generic;
 using BcfToolkit.Builder.Bcf21.Interfaces;
 using BcfToolkit.Builder.Interfaces;
 using BcfToolkit.Model;
@@ -8,11 +7,10 @@ using BcfToolkit.Model.Bcf21;
 namespace BcfToolkit.Builder.Bcf21;
 
 public partial class BcfBuilder : IBcfBuilder<
-  BcfBuilder,
-  MarkupBuilder,
-  ProjectBuilder>,
+    BcfBuilder,
+    MarkupBuilder,
+    ProjectBuilder>,
   IDefaultBuilder<BcfBuilder> {
-
   private readonly Bcf _bcf = new();
 
   public BcfBuilder AddMarkup(Action<MarkupBuilder> builder) {
@@ -22,25 +20,20 @@ public partial class BcfBuilder : IBcfBuilder<
     return this;
   }
 
-  public BcfBuilder AddMarkups(List<Markup> markups) {
-    markups.ForEach(m => _bcf.Markups.Add(m));
-    return this;
-  }
-
   public BcfBuilder SetProject(Action<ProjectBuilder> builder) {
     var project =
-      (ProjectExtension)BuilderUtils.BuildItem<ProjectBuilder, IProject>(builder);
+      (ProjectExtension)BuilderUtils.BuildItem<ProjectBuilder, IProject>(
+        builder);
     _bcf.Project = project;
-    return this;
-  }
-
-  public BcfBuilder WithDefaults() {
-    this
-      .AddMarkup(m => m.WithDefaults());
     return this;
   }
 
   public Bcf Build() {
     return BuilderUtils.ValidateItem(_bcf);
+  }
+
+  public BcfBuilder WithDefaults() {
+    AddMarkup(m => m.WithDefaults());
+    return this;
   }
 }
