+++ src/bcf-toolkit/Builder/Bcf21/BcfBuilderExtensions.cs
@@ -1,3 +1,4 @@
+using System.Collections.Generic;
 using System.IO;
 using System.Threading.Tasks;
 using BcfToolkit.Model.Bcf21;
@@ -6,8 +7,14 @@ using BcfToolkit.Utils;
 namespace BcfToolkit.Builder.Bcf21;
 
 public partial class BcfBuilder {
+  public BcfBuilder AddMarkups(List<Markup> markups) {
+    markups.ForEach(m => _bcf.Markups.Add(m));
+    return this;
+  }
+
   public async Task<Bcf> BuildFromStream(Stream source) {
-    _bcf.Markups = await BcfExtensions.ParseMarkups<Markup, VisualizationInfo>(source);
+    _bcf.Markups =
+      await BcfExtensions.ParseMarkups<Markup, VisualizationInfo>(source);
     _bcf.Project = await BcfExtensions.ParseProject<ProjectExtension>(source);
     return BuilderUtils.ValidateItem(_bcf);
   }
